+++ frontend/src/pages/Profile/ProfileEntities.tsx
@@ -3,17 +3,17 @@ import { Grid, Typography } from "@mui/material";
 import { ProfileDto } from "../../types/profile/profile.dto";
 
 interface ProfileEntities {
-	profile: ProfileDto | null;
+	profile: ProfileDto;
 }
 
 export const ProfileEntities: React.FC<ProfileEntities> = ({ profile }) => {
 	return (
 		<Grid>
-			<Typography variant="h6">{profile?.email}</Typography>
-			<Typography variant="h6">{profile?.email}</Typography>
-			<Typography variant="h6">{profile?.email}</Typography>
-			<Typography variant="h6">{profile?.email}</Typography>
-			<Typography variant="h6">{profile?.email}</Typography>
+			<Typography variant="h6">{profile.email}</Typography>
+			<Typography variant="h6">{profile.email}</Typography>
+			<Typography variant="h6">{profile.email}</Typography>
+			<Typography variant="h6">{profile.email}</Typography>
+			<Typography variant="h6">{profile.email}</Typography>
 		</Grid>
 	);
 };